fix(schema): emit double, not float, for LinkML Float mapping (KI-068)

+++ src/ontolith/schema/linkml.py
@@ -41,7 +41,19 @@ _ValueType = Literal["Text", "Integer", "Float", "Boolean", "Date", "DateTime", 
 _VALUE_TYPE_TO_LINKML_RANGE: dict[str, str] = {
     "Text": "string",
     "Integer": "integer",
-    "Float": "float",
+    # "double" (64-bit), not "float" (32-bit) - Ontolith's Float is backed by
+    # Python's float/IEEE-754 double throughout (identical to what schema/rdf.py's
+    # RDF/OWL bridge already emits, XSD.double). Ontolith's own IR -> YAML -> IR
+    # round-trip was never lossy either way (the reverse table below has always
+    # accepted both spellings) - the defect this fixes is a mis-declared
+    # precision that only matters to a downstream LinkML-consuming tool (codegen,
+    # SHACL/JSON-Schema generation) resolving `float` to 32-bit xsd:float, not
+    # anything Ontolith itself round-trips through. (KI-068; ADR-0013 Update -
+    # ADR-0036's own Consequences section disclosed the mismatch and declined to
+    # reconcile it as out of that ADR's scope, not "named it as future work".)
+    # The reverse mapping already accepted "double" before this change, so this
+    # makes the two tables consistent with each other, not just internally.
+    "Float": "double",
     "Boolean": "boolean",
     "Date": "date",
     "DateTime": "datetime",
@@ -52,6 +64,11 @@ _VALUE_TYPE_TO_LINKML_RANGE: dict[str, str] = {
 _LINKML_RANGE_TO_VALUE_TYPE: dict[str, str] = {
     "string": "Text",
     "integer": "Integer",
+    # "float" stays accepted on import (not just "double") for backward
+    # compatibility with hand-authored LinkML and any schema exported by an
+    # older Ontolith version before the emission side switched to "double"
+    # above (KI-068) - both losslessly represent Ontolith's own Float either
+    # way, since Python float parsing doesn't distinguish the two on read.
     "float": "Float",
     "double": "Float",
     "boolean": "Boolean",
